♻️ refactor(mq-lang): split tarn.rs into focused submodules
tarn.rs had grown to ~3300 lines mixing engine orchestration with
bytecode-cache handling, nodes-split AST helpers, bytecode
disassembly, and its own ~1800-line test module. Pull each concern
into its own file under tarn/: cache.rs (repeated-eval bytecode
caching), nodes_split.rs (splitting a program around `nodes`),
disasm.rs (bytecode pretty-printing for --dump-bytecode), and
tests.rs (the existing #[cfg(test)] mod, moved as-is).

tarn.rs itself now holds only the front door: Error, VmState/TarnVm,
and the session/run orchestration that stays entangled with
#[cfg(feature = "debugger")] wiring. No behavior change; verified
with cargo build/clippy across --no-default-features, --features
tarn, and --features debugger, plus just test-all (9937 passed) and
just test-mq-vm.
